Fixed: corner grab in touch-aware browsers such as recent Chrome.
Without this fix, turn.js would detect that a browser was touch-enabled and if it found it was, not react to mouse events. Modern versions of Chrome are touch-enabled, but that doesn't mean there is no mouse. Furthermore, a device could have both a touch device and a mouse.

This fix makes sure that touch enabled browsers still react to mouse events.